requirements.txt
Django==4.2.3
django-crispy-forms==1.13.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
Pillow==10.0.0
